sosreport: fix name matching of the generated report file
In src/kimchi/model.py, function _sosreport_generate() collects the
generated debug report through pattern matching against the file name.
In the current implementation, it actually collects the last file in the
matching file set. This works most of the time but is not correct.

This patch fixes the problem, skips on *.md5 file, collects the correct
file, and raises exception when no file can be found.

This patch also renames "fi" to "reportFile" to improve readability and
ease of searching.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>