Tests: Generate UUID and use as temp report file name
test_debug_reports runs the debug report tool to generate report file.
It's possible that the tool runs too long and timeout expires. While the
test is over, the tool continues to run and generate report file. The
leftover report file causes conflict with report file name used in the
the next run of test_debug_reports.

This patch generates UUID for each run and use it as the part of file
name. It also deletes all leftover files generated from previous runs.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>