storagepool: rename and consolidate arguments of creating (front-end)
As we are adding support to new type of storage pool, the current naming
scheme of the storage pool creating arguments should be rearranged to be
more extendable. This patch renames some arguments and consolidates the
argument of the same purposes as follow.

nfsserver -> source.host
nfspath -> source.path
devices -> source.devices

This patch contains the front end changes of the consolidation work. It
just changes the formData transferred to kimchi backend, and leaves the
templates and other related scripts untouched.

v2 -> v5
  No changes

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>