Add jquery widget dropDown button
v1-v2 Modify the widget with create and destroy function
and drop redundant function.
This patch is working for make dropDown button a jquery
widget.
We can create a dropdown button simply with:
$('#buttonId').dropdownButton();

And in html:

<div id="buttonId">
    <span class="text">$_("Button label")</span><span class="arrow"></span>
    <div>
       <button><span class="text">$_("dropdownbutton1")</span></button>
       <button><span class="text">$_("dropdownbutton2")</span></button>
    </div>
</div>

If we need some other style, you can add it to html, and bind action in js.
Not a perfect widget, but a beginning.
Signed-off-by: zhoumeina <zhoumein@linux.vnet.ibm.com>