Support VM template for Power machine
VM on power machine requires different parameters from X86 machine.
So kimchi should provide different template according to the architecture.
IDE bus is not available on power platform. So change the device
bus of cdrom to scsi instead of ide for VM on power. The same
problem exists for old nic model, and change it to rtl8139 instead.

Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>