UI bug fix: Set vlan_id value only if 'Enable VLAN' is checked
Commit 2f54aaba added a checkbox to provide choice of enabling vlan tagging
while creating a bridged network, but it doesn't check the checkbox before
setting the vlan_id to send server request.

Because that when the user tries to create a bridged network without enabling
VLAN the following error is raised:
Invalid parameter: 'None is not of type u'integer''

Fix it by setting vlan_id only if the vlan checkbox is checked.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>