Add the iSCSI dependency packages to README
The README file contains instructions to build Kimchi from the source
tree, including the exact commands (e.g. yum, apt-get, zypper) to install the
dependencies on the main Linux distros. But none of those commands contain the
iSCSI package which is needed to validate some types of storage pool devices.

Add the iSCSI package ("iscsi-initiator-utils" or "open-iscsi", depending on the
distro) to the install commands in the README file.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>