storagepool: Use callback functions when deleting a pool
After deleting a storage pool, the page may update the pool list even if
the operation has not been finished. That results in inconsistent
information in the UI, like duplicate or missing entries. Also, if there
is an error during this operation, nothing is displayed to the user.

Make sure the storage pool list is updated only after the delete
operation finishes successfully, and display an error message if
something goes wrong.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>