Remove 'encoding=iso-8859-1' from json.dumps()
The encoding option for json.dumps() is used to indicate the charsets
of the passed in str. Currently,  we use the encoding of iso-8859-1,
but actually we don't have any data encoded in iso-8859-1. It doesn't
cause any problem because the strings in the data passed are ascii chars
or unicode. If we pass a utf-8 str, it will cause an expected result.
So we could remove the option, and after that it can accept utf-8 str too
because json.dumps use utf-8 as default encoding.

Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>