Host's software update: Update backend.
Update model and mockmodel to support backend opertions.
Add new file implementing backend operations, with four new classes:

1) SoftwareUpdate (object): Class to represent and operate with OS software
update system in Kimchi's perspective. It's agnostic to host's package management
system, and can execute all operations necessary: get all packages to update,
get information about one package and execute the update. This class will load
in runtime the necessary classes to work with the host's package management:
YumUpdate for YUM systems based, AptUpdate for APT systems based and ZypperUpdate
for Zypper systems based.

2) YumUpdate (object): Class to represent and operate with YUM. Loaded only on
those systems that supports YUM, it's responsible to connect and collect
information of the packages to be updated. Also it's responsible to execute the
update of the system.

3) AptUpdate (object): Class to represent and operate with APT. Loaded only on
those systems that supports APT, it's responsible to connect and collect
information of the packages to be updated. Also it's responsible to execute the
update of the system.

4) ZypperUpdate (object): Class to represent and operate with Zypper. Loaded only
on those systems that supports Zypper, it's responsible to connect and collect
information of the packages to be updated. Also it's responsible to execute the
update of the system.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>
Signed-off-by: Ramon Medeiros <ramonn@linux.vnet.ibm.com>
Signed-off-by: Aline Manera <alinefm@br.ibm.com>