Return 403 Forbidden when user does not have permission to access kimchi
We need to differ when user is not logged into kimchi and when he does not have
enough permissions to perform an operation.
When user is not logged in, the UI needs to know to ask him to login
again into kimchi. In this case, return 401 Unauthorized
When user does not have enough permission returns 403 Forbidden

401 Unauthorized - The request requires user authentication.
403 Forbidden - The server understood the request, but is refusing to
fulfill it.

For reference: http://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html

Also, updated the tests cases to reflect those changes and po files as
new messages were added.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>