Github bug #326: run_command: killing all children processes
The subprocess.kill() call does not handle well the created children
of the process created by the Popen() call. In some cases (for
example, the mount command) the parent processes creates children
processes that aren't killed, leaving the parent thread hanging out
in an infinite loop waiting for kill() to finish.

An attempt was made using the os.killpg() call, but to no success.
Using psutils facilities to kill all the children before killing the
opened process was easier and clearer than the alternatives (opening
a new thread and trying to kill the thread, for example).

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>