Fix debug report naming problem (UI)
The tool sosreport only accepts letters, digits and hyphen in the name
of the report, non alphanumeric characteres are removed from the file
name and then Kimchi is not able to find the file.

This patch modifies the UI html and js in order to show the right
information and catch the right return error message.

Implements name checking in frontend.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>