issue #325: Use RamSession instead of FileSession
There is a readers/writes problem in cherry.sessions.FileSession
implementation which makes it non thread safe.

From https://groups.google.com/forum/#!topic/cherrypy-users/biitlom41T8
"When the session tool is turned on it hooks the lib.session.save
function onto 'before_finalize' so that the session data is always
re-saved at the end of each request.  For RamSession this is fine
because save just overwrites a dictionary value.  FileSession uses
pickle.dump to write the session data to file and uses pickle.load to
load it. If two requests overlap in just the wrong way then it's
possible that pickle.load gets called halfway through the pickle.dump
which throws an EOFError. More concurrent requests will increase the
chance of this happening."

For reference: http://www.stderr.org/doc/cherrypy-doc/html/howto/node17.html

On Host tab, more concurrent requests happen, specially while creating a
debug report: /host/stats and /tasks/<id>. Because that the problem was
identified there.

To solve this issue, there are 2 options:
1) overwrite cherrypy.session.FileSession._load to fix the
readers/writers and make it thread safe;
or
2) Use RamSession instead of FileSession

As using RamSession makes the application faster because it will not need to
read/write file for each request, this is the best solution in this case.

This patch also uses acquire_lock() and release_lock() for each
read/write session data occurrence.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>