Github bug #327: NFS pool workaround: model changes
Implementing a workaround in '_get_storagepool_vols_num',
'lookup', 'activate' and 'deactivate' methods to avoid
libvirt calls in NFS pools when the corresponding NFS server
is unreachable.

In the 'lookup' method the state of the pool returned in
such scenario is 'inaccessible' to warn the UI that the pool
isn't available for normal operation.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>