Fix screenshots and debug reports paths.
The definitions of the screenshots and debug reports paths in Cherrypy
config were pointing to a relative path that, when concatenated with the
root path defined in the [/] section, were not mapping to the correct path
when Kimchi is installed through its RPM package.

This fix removes the use of the root path in the [/] config section and
specifies the absolute path for all resources. That way we can easily
handle running Kimchi directly from the git source tree or after a
package installation --- in the later case the paths will not always be
relative to the same root path, but will rather be spread in the file system
according to the unix file system standard.

Signed-off-by: Leonardo Garcia <lagarcia@br.ibm.com>