Fix debug report UI: Error handling and line selection
If a user tries to delete a report and some error occurs, no message is
showed in the UI. Also, after delete the report, all lines become
unselected but remove and download buttons are enabled. So user is able
to click on remove button again, but will not know which report is going
to be deleted. This patch disables the buttons, so user have to select
the report again to perform the operations, avoiding mistakes.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>