bug fix: Enable NFS path field when a NFS server is provided
While creating a NFS storage pool, the NFS path field was not editable
after entering "NFS server IP" the NFS path, unless there was a
mouse click on other part of the window or use "tab".

It is because .change() has different behaviors according to element
it is linked.
The change event is sent to an element when its value changes.
But for some element types the event is deferred until the element loses
focus - what was happening with "NFS server IP" field.

So use .keyup() to listen when user types in the field to fix this issue.
The keyup event is sent to an element when the user releases a key on
the keyboard.

For reference:
  - http://api.jquery.com/change/
  - http://api.jquery.com/keyup/

Signed-off-by: Aline Manera <alinefm@br.ibm.com>