kimchi.control.utils: fix request parsing in generate_action_handler()
generate_action_handler() parses the request and collects the action
arguments. The problem is it calls parse_request() in a generation
expression. This implies parse_request() is called for each action
argument. However parse_request() should be called only once, because it
reads over the request body and turns it to a dict. When it is called a
second time, the request body is already over, so parse_request() fails.

This patch stores the result from parse_request() in a variable, and
uses the variable in the generation expression.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>