While trying to extend a logical pool I got the following error message:
"/dev/vdb is not a valid disk/partition. Could not add it to the pool
my-logical-pool"

But the same partition "/dev/vdb" is listed on GET /host/partitions.
Then I realized the "blkid" is not the correct command to check a
partition exists for raw disks.

From blkid man page:
It can determine the type of content (e.g. filesystem or swap) that a
block
device holds,  and  also  attributes  (tokens, NAME=value pairs) from
the
content metadata (e.g. LABEL or UUID fields).

Which is not useful for a raw disk.

From lsblk man page:
lsblk  lists  information  about  all  or  the specified block devices.

So use "lsblk" command instead of "blkid" to verify the disk exists in
the system.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>